Re-enable additional scrollable fetch test
